refactor: separate axios functions and inject in registration component
