+++ src/core/models/CreateUserResponse.ts
@@ -1,4 +1,4 @@
-interface CreateUserResponse {
+export interface CreateUserResponse {
     name: string,
     email: string,
     password: string,
